fix(formats-parquet): der Coverage-Gate-Befund schliesst das Modul UND zwei Defekte, die er aufdeckte
`adapters/driven/formats-parquet` hatte weder eine eigene 90-%-Regel noch
stand es im Wurzel-Aggregat -- 22 Produktivdateien von keiner Seite
gedeckt. Gemessen mit einer temporaeren minBound(100)-Regel (kover nennt
beim Fehlschlag die reale Zahl): 77,2 %. Kein strukturelles Hindernis --
die Luecke lag in reinen JVM-Koerzierungs- und Type-Dispatch-Funktionen,
deren NATUeRLICHER JDBC-Typ pro Spalte laengst getestet war, nicht die
anderen Java-Typen, die ein Treiber ebenso liefert, und nicht die
selteneren NeutralType-Varianten (nur 4-6 von 21 liefen je durch einen
Manifest-Rundlauf).

Das systematische Testen fand dabei zwei echte, zuvor unbekannte Defekte:

- Eine Time-Spalte liess sich gar nicht als Parquet schreiben. TIME(MICROS)
  verlangt physisch INT64 (Parquet-Spezifikation); der Code deklarierte
  INT32 und scheiterte beim Schema-Bau, vor der ersten Zeile. Kein Randfall:
  Mikrosekunden seit Mitternacht (86 400 000 000) ueberschreiten Int.MAX_VALUE
  ohnehin. Kein Test hatte je eine Time-Spalte durch den echten Schreibpfad
  gefuehrt. Vier Stellen korrigiert (Schema-Bau, Writer, Reader, Rueck-
  Umsetzer MessageType->NeutralType), die laut eigener KDoc "zusammen
  geaendert werden muessen" -- die KDoc kannte das Risiko, der Fehler war
  trotzdem da.
- Eine FullText-Spalte liess sich in ein Bundle-Manifest schreiben, aber
  nicht mehr lesen -- der Schreiber kannte den Kind-String, der Leser nicht,
  Absturz beim naechsten Lesen des eigenen Manifests. Gefunden durch einen
  Rundlauf-Test ueber alle 21 NeutralType-Varianten statt der bis dahin
  ueblichen vier bis sechs.

Neue Tests: ParquetSingleFileResolverTest (zwei komplett ungetestete
Klassen, 0 % -> 100 %), ParquetGroupValueWriterTest (Koerzierungszweige
ueber einen echten SimpleGroupFactory-Rumpf, nicht gemockt),
ManifestNeutralTypeRoundTripTest (alle 21 Varianten), plus eine Time-Spalte
im bestehenden Mixed-Schema-Rundlauf.

Endergebnis: 90,3 % Zeilen-Deckung (vorher 77,2 %). Modul traegt jetzt
minBound(90) und steht im Wurzel-Aggregat.

Was offen blieb: die Mechanisierung des Modul-Abgleichs (Regel x Aggregat x
Ledger) selbst -- kein eigenes Skript gebaut, bleibt manuelle Pruefung beim
naechsten Release.

Gruen: alle 49 Module (Vollbau, root build.gradle.kts geaendert), a-check,
solid-suppression-gate, docs-check.

+++ adapters/driven/formats-parquet/src/main/kotlin/dev/dmigrate/format/parquet/ChunkSchemaToParquetMessageType.kt
@@ -77,7 +77,14 @@ internal object ChunkSchemaToParquetMessageType {
                 .`as`(LogicalTypeAnnotation.dateType())
                 .named(column.name)
 
-            is NeutralType.Time -> Types.primitive(PrimitiveTypeName.INT32, repetition)
+            // TIME(unit=MICROS) verlangt laut Parquet-Spezifikation INT64 als
+            // physischen Typ -- Types.primitive() wirft sonst beim Bauen des
+            // MessageType. Nachgemessen 2026-09-13 beim Schliessen einer
+            // Coverage-Luecke: kein bestehender Test schrieb je eine
+            // Time-Spalte, der Fehler war deshalb latent. Gehoert zusammen
+            // mit ParquetGroupValueWriter/-Reader und
+            // ParquetMessageTypeToChunkSchema, die ebenfalls INT64 erwarten.
+            is NeutralType.Time -> Types.primitive(PrimitiveTypeName.INT64, repetition)
                 .`as`(LogicalTypeAnnotation.timeType(/* isAdjustedToUTC = */ false, MICROS))
                 .named(column.name)
